Support custom external commands
1. TST custom commands are scripts that can be run from within
the tst commander and that can interact with results of other tst
commands. To add a command named <command>, the user must add a
script named tst-<command> to the ~/.tst/commands directory and
set the TST_CUSTOM_COMMANDS environment variable with the name of
the commands separated by colons.

The installation script was changed to overwrite an existing
installation so that scripts of custom commands are not deleted
in a new installation.