Fix bug tstlib.py and improve status and checkout
Fix index reported by activity_differ. The algorithm used to
calculate the line index of each changed line was wrong, because
difflib.ndiff does not alternate removed and added lines as I
expected.

The checkout (still checkoute) now warns the user if he tries to
checkout over an existing and changed activity. It requires the
use of option --overwrite.

The output of the status command had minor improvements. In
particular, it puts the name of the activity in the top of the
report.