Add experimental checkout and commit activities
An initial implementation of four new commands is provided in
this commit: new, checkout2, status and commit2. Observe that
I named the commands checkout2 and commit2 to avoid mess with
existing functionality of checkout and commit for assignments.

In future, I plan to integrate them both. Now, however, they will
serve only to allow me to proceed with a proof of concept of the
general idea.