Replace get_context with read_specification
The new function reads the specification of the current
directory. If a specific file name is given, that file is
read. Otherwise, the function reads the tst.yaml or the
tst.json files. If these files don't exist, the function
looks for any yaml file and, then, any json files. If only
one exists, it is taken as the specification file.
Otherwise, the function exits with a message to the user.