Add new dispatcher support
The new dispatcher design is based on the support of
setuptools for console scripts and entry points. Commands
now will be dispatched by the tst/commands/dispatcher.py
script. Internal commands will be kept in separate modules
within the same directory and will be imported by the
dispatcher when necessary.