Add tst object fetch and saving support to tst
Now a fully functional get method has been added to the Site
class. The method fetches an object from a site, given the
appropriate key. Given the module uses requests and cachecontrol,
all requests are cached in the ~/.tst/cache directory.

A function save_files is also added to the tst module. It allows
saving a list of files into a given directory. If the files
already exist they are overwritten (a message can be printed to
the console).

- minor improvement of the get_site function
- function save2fs renamed to save_files
- helper functions: validate_tst_object, save_file