Add working implementation of checkout command
The checkout command interprets the command line arguments, and
identify the Site, the key and the directory to be used. Once
this is done, the files are saved using tst.save_files. If the
files already exist, a confirmation is required from the user,
except if he has used either the --yes or -y options. The
command also saves the .tst/assignment.json file containing
metadata about the object in the directory. This saving is
performed silently.

Minor changes:
- improve the help message
- add helper functions: process_args, existing_files
- improve the parse_cli_args helper function