Make checkout work with activities in directories
Now if the url created from the activity key doesn't provide a
valid json, the url is assumed to be a directory. As such, the
first resource fetched will be `$site_url/key/tst.yaml`. Also,
all other files mentioned in this file will be fetched, unless
it is already available in the `files` property in `tst.yaml`.