Add version number control, command and option
The version number is now maintained in file VERSION. It is read
by the setup.py. The command `version` was also added and the
`--version` and `-v` options are do exactly the same. When the
first argument is one of them, the version number is read using
the pkg_resources module. This is the 4th technique mentioned in
https://packaging.python.org/guides/single-sourcing-package-version/