Add http post method to Site + refactor login cmd
This commit moves the post method to the Site class within the
tst.py module. The method is a candidate to become a generic http
post facility method to help configure every request to the site
in a single place. It adds headers and cookies as required. It
still requires that client code imports requests errors to deal
with them.