Refactor tst.js use of requests session
Add a get_session method to Site. This method now is the only one
to create and configure session objects from the requests
package. This allows to have a single point of configuration for
headers, cookies and cache control. The rest of the changes were
simply adaptations to this refactoring