Improve script test specification format
Script tests can be specified with a simple command property,
now. The command will be run as is. If the the file of the answer
being tested must be added to the command, the command should
include a `{}`. The type property is now optional for both script
and io tests. It is inferred from the properties used.