Rewrite tst-test user interface
Replace the awkward report_results function with the Reporter
hierarchy class: Reporter, FilterReporter, WorkerReporter and
DebugReporter. Also add the ColorizingStreamHandler to deal with
colored logs to the console. This class is very likely to be
moved to a more generic module when the logging solution is
implemented in all modules.