Fix checkout requests sequence
The original sequence of requests was wrong. It used to
fetch the activity ID and then it would consider the possibility
that the ID was a directory and fetch a tst.yaml file within
it. Howeverm it used to try the second alternative even when
the reason given by the first fetch was not 404. This commit
fixes such behavior.