improve external diff support
External diff tool and arguments can now be specified in the
configuration file (property `diff`). If it is not configured,
the default tool used is simply `diff`. The command creates or
overwrites two files in the `/tmp` directory, namely:
`/tmp/output` and `/tmp/stdout`. The former will contain the
output expected in the test, while the latter will contain the
actual output produced by the subject under test. See that if
multiple tests fail, the final contents will correspond to the
last test case.