fix: json report not printed if file doesnt exist
First, the json report didn't print a json output when a single
filename is given in command line and the file doesn't exist. If
more than one file is given, however, the json output is
correctly printed. Second, the summary for the error (when the
file doesn't exist) is more specific.